% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.mediate.mcontiyconti.or.R
\name{hetero.mediate.mcontiyconti}
\alias{hetero.mediate.mcontiyconti}
\title{Causal Mediation Analysis for Heterogeneous Effect (Continuous Mediator and Continuous Outcome)}
\usage{
hetero.mediate.mcontiyconti(
  model.m,
  model.y,
  modifier,
  treat,
  mediator,
  n,
  conf.level = 0.95
)
}
\arguments{
\item{model.m}{A fitted model object for mediator.}

\item{model.y}{A fitted model object for outcome.}

\item{modifier}{A character string indicating the name of the modifier used in the models.}

\item{treat}{A character string indicating the name of the treatment indicator used in the models.}

\item{mediator}{A character string indicating the name of the mediator used in the models.}

\item{n}{Sample size.}

\item{conf.level}{Level of the returned two-sided confidence intervals. Default is set to 0.95, which returns the 2.5 and 97.5 percentiles of the quantities.}
}
\value{
The estimate, standard error, p value and confidence interval of natural direct and indirect effect heterogeneity, pure direct and indirect effect heterogeneity, mediated interactive effect heterogeneity, controlled direct effect heterogeneity, and total effect heterogeneity on the risk difference scale.
}
\description{
This function is for the estimation of various causal heterogeneous effects for causal mediation analysis, including natural direct and indirect effect heterogeneity, pure direct and indirect effect heterogeneity, mediated interactive effect heterogeneity, controlled direct effect heterogeneity, and total effect heterogeneity. Both the mediator and the outcome are continuous.
}
