% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.mediate.R
\name{hetero.mediate}
\alias{hetero.mediate}
\title{Causal Mediation Analysis for Heterogeneous Effect}
\usage{
hetero.mediate(
  model.m,
  model.y,
  modifier,
  treat,
  mediator,
  n,
  conf.level = 0.95,
  mreg = "linear",
  yreg = "linear"
)
}
\arguments{
\item{model.m}{A fitted model object for mediator.}

\item{model.y}{A fitted model object for outcome.}

\item{modifier}{A character string indicating the name of the modifier used in the models.}

\item{treat}{A character string indicating the name of the treatment indicator used in the models.}

\item{mediator}{A character string indicating the name of the mediator used in the models.}

\item{n}{Sample size.}

\item{conf.level}{Level of the returned two-sided confidence intervals. Default is set to 0.95, which returns the 2.5 and 97.5 percentiles of the quantities.}

\item{mreg}{Type of mediator model. Can be of class 'linear', 'logistic', 'loglinear'. Default is 'linear'.}

\item{yreg}{Type of outcome model. Can be of class 'linear', 'logistic', 'loglinear'. Default is 'linear'.}
}
\value{
The estimate, standard error, p value and confidence interval of natural direct and indirect effect heterogeneity, pure direct and indirect effect heterogeneity, mediated interactive effect heterogeneity, controlled direct effect heterogeneity, and total effect heterogeneity.
}
\description{
This function is for the estimation of various causal heterogeneous effects for causal mediation analysis, including natural direct and indirect effect heterogeneity, pure direct and indirect effect heterogeneity, mediated interactive effect heterogeneity, controlled direct effect heterogeneity, and total effect heterogeneity.
}
